move jar deps for unpacked jars calculation test to the unpacked jars file
