+++ tests/python/pants_test/backend/jvm/targets/test_unpacked_jars.py
@@ -4,14 +4,26 @@
 
 from __future__ import absolute_import, division, print_function, unicode_literals
 
+from textwrap import dedent
+
 from pants.backend.jvm.targets.import_jars_mixin import ImportJarsMixin
 from pants.backend.jvm.targets.jar_library import JarLibrary
 from pants.backend.jvm.targets.unpacked_jars import UnpackedJars
+from pants.build_graph.build_file_aliases import BuildFileAliases
 from pants.java.jar.jar_dependency import JarDependency
 from pants_test.test_base import TestBase
 
 
 class UnpackedJarsTest(TestBase):
+
+  @classmethod
+  def alias_groups(cls):
+    return BuildFileAliases(
+      targets={
+        'jar_library': JarLibrary,
+        'unpacked_jars': UnpackedJars,
+      },
+      objects={'jar': JarDependency})
 
   def test_empty_libraries(self):
     with self.assertRaises(UnpackedJars.ExpectedLibrariesError):
@@ -34,3 +46,45 @@ class UnpackedJarsTest(TestBase):
     target = self.make_target(':foo', UnpackedJars, libraries=[':wrong-type'])
     with self.assertRaises(ImportJarsMixin.WrongTargetTypeError):
       target.imported_targets
+
+  def test_has_all_imported_jar_deps(self):
+    def assert_dep(dep, org, name, rev):
+      self.assertTrue(isinstance(dep, JarDependency))
+      self.assertEqual(org, dep.org)
+      self.assertEqual(name, dep.name)
+      self.assertEqual(rev, dep.rev)
+
+    self.add_to_build_file('BUILD', dedent('''
+    jar_library(name='lib1',
+      jars=[
+        jar(org='testOrg1', name='testName1', rev='123'),
+      ],
+    )
+    jar_library(name='lib2',
+      jars=[
+        jar(org='testOrg2', name='testName2', rev='456'),
+        jar(org='testOrg3', name='testName3', rev='789'),
+      ],
+    )
+    unpacked_jars(name='unpacked-lib',
+      libraries=[':lib1', ':lib2'],
+    )
+    '''))
+    lib1 = self.target('//:lib1')
+    self.assertIsInstance(lib1, JarLibrary)
+    self.assertEqual(1, len(lib1.jar_dependencies))
+    assert_dep(lib1.jar_dependencies[0], 'testOrg1', 'testName1', '123')
+
+    lib2 = self.target('//:lib2')
+    self.assertIsInstance(lib2, JarLibrary)
+    self.assertEqual(2, len(lib2.jar_dependencies))
+    assert_dep(lib2.jar_dependencies[0], 'testOrg2', 'testName2', '456')
+    assert_dep(lib2.jar_dependencies[1], 'testOrg3', 'testName3', '789')
+
+    unpacked_lib = self.target('//:unpacked-lib')
+    unpacked_jar_deps = unpacked_lib.all_imported_jar_deps
+
+    self.assertEqual(3, len(unpacked_jar_deps))
+    assert_dep(lib1.jar_dependencies[0], 'testOrg1', 'testName1', '123')
+    assert_dep(lib2.jar_dependencies[0], 'testOrg2', 'testName2', '456')
+    assert_dep(lib2.jar_dependencies[1], 'testOrg3', 'testName3', '789')
